fixed up unit tests and removed random files
