--- TestApp/Views/Home/Index.cshtml
@@ -1,9 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Home Page";
-}
-@inject NormalClass normalClass
-@inject ITest Class1 "NumberOne"
-<div class="text-center">
-    <h1 class="display-4">Welcome</h1>
-    <p>Learn about <a href="https://learn.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
-</div>
