--- TestApp/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using TestApp
-@using TestApp.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
